+++ src/main/java/com/b1n_ry/yigd/compat/InvModCompat.java
@@ -3,6 +3,7 @@ package com.b1n_ry.yigd.compat;
 import com.b1n_ry.yigd.compat.misc_compat_mods.CommonProtectionApiCompat;
 import com.b1n_ry.yigd.compat.misc_compat_mods.OrpheusCompat;
 import com.b1n_ry.yigd.compat.misc_compat_mods.RespawnObelisksCompat;
+import com.b1n_ry.yigd.compat.misc_compat_mods.TwilightCompat;
 import com.b1n_ry.yigd.config.YigdConfig;
 import net.fabricmc.loader.api.FabricLoader;
 import net.minecraft.nbt.NbtCompound;
@@ -51,6 +52,8 @@ public interface InvModCompat<T> {
             OrpheusCompat.init();
         if (compatConfig.enableRespawnObelisksCompat && loader.isModLoaded("respawnobelisks"))
             RespawnObelisksCompat.init();
+        if (loader.isModLoaded("twilightforest"))
+            TwilightCompat.init();
     }
 
     String getModName();